@@ -1,4 +1,4 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 __title__ = "browserbase"
-__version__ = "1.4.0"  # x-release-please-version
+__version__ = "1.5.0-alpha.0"  # x-release-please-version
