@@ -16,9 +16,9 @@ from .._response import (
     async_to_streamed_response_wrapper,
 )
 from .._base_client import make_request_options
-from ..types.context import Context
 from ..types.context_create_response import ContextCreateResponse
 from ..types.context_update_response import ContextUpdateResponse
+from ..types.context_retrieve_response import ContextRetrieveResponse
 
 __all__ = ["ContextsResource", "AsyncContextsResource"]
 
@@ -89,9 +89,9 @@ class ContextsResource(SyncAPIResource):
         extra_query: Query | None = None,
         extra_body: Body | None = None,
         timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
-    ) -> Context:
+    ) -> ContextRetrieveResponse:
         """
-        Context
+        Get a Context
 
         Args:
           extra_headers: Send extra headers
@@ -109,7 +109,7 @@ class ContextsResource(SyncAPIResource):
             options=make_request_options(
                 extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
             ),
-            cast_to=Context,
+            cast_to=ContextRetrieveResponse,
         )
 
     def update(
@@ -124,7 +124,7 @@ class ContextsResource(SyncAPIResource):
         timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
     ) -> ContextUpdateResponse:
         """
-        Update Context
+        Update a Context
 
         Args:
           extra_headers: Send extra headers
@@ -212,9 +212,9 @@ class AsyncContextsResource(AsyncAPIResource):
         extra_query: Query | None = None,
         extra_body: Body | None = None,
         timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
-    ) -> Context:
+    ) -> ContextRetrieveResponse:
         """
-        Context
+        Get a Context
 
         Args:
           extra_headers: Send extra headers
@@ -232,7 +232,7 @@ class AsyncContextsResource(AsyncAPIResource):
             options=make_request_options(
                 extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
             ),
-            cast_to=Context,
+            cast_to=ContextRetrieveResponse,
         )
 
     async def update(
@@ -247,7 +247,7 @@ class AsyncContextsResource(AsyncAPIResource):
         timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
     ) -> ContextUpdateResponse:
         """
-        Update Context
+        Update a Context
 
         Args:
           extra_headers: Send extra headers
