@@ -6,21 +6,17 @@ from pydantic import Field as FieldInfo
 
 from .._models import BaseModel
 
-__all__ = ["Project"]
+__all__ = ["ExtensionRetrieveResponse"]
 
 
-class Project(BaseModel):
+class ExtensionRetrieveResponse(BaseModel):
     id: str
-
-    concurrency: int
-    """The maximum number of sessions that this project can run concurrently."""
 
     created_at: datetime = FieldInfo(alias="createdAt")
 
-    default_timeout: int = FieldInfo(alias="defaultTimeout")
+    file_name: str = FieldInfo(alias="fileName")
 
-    name: str
-
-    owner_id: str = FieldInfo(alias="ownerId")
+    project_id: str = FieldInfo(alias="projectId")
+    """The Project ID linked to the uploaded Extension."""
 
     updated_at: datetime = FieldInfo(alias="updatedAt")
