@@ -4,10 +4,10 @@ from pydantic import Field as FieldInfo
 
 from .._models import BaseModel
 
-__all__ = ["ProjectUsage"]
+__all__ = ["ProjectUsageResponse"]
 
 
-class ProjectUsage(BaseModel):
+class ProjectUsageResponse(BaseModel):
     browser_minutes: int = FieldInfo(alias="browserMinutes")
 
     proxy_bytes: int = FieldInfo(alias="proxyBytes")
