@@ -10,10 +10,10 @@ import pytest
 from browserbase import Browserbase, AsyncBrowserbase
 from tests.utils import assert_matches_type
 from browserbase.types import (
-    Session,
-    SessionLiveURLs,
     SessionListResponse,
+    SessionDebugResponse,
     SessionCreateResponse,
+    SessionUpdateResponse,
     SessionRetrieveResponse,
 )
 
@@ -57,7 +57,9 @@ class TestSessions:
                         "min_width": 0,
                     },
                 },
+                "headful": True,
                 "log_session": True,
+                "os": "windows",
                 "record_session": True,
                 "solve_captchas": True,
                 "viewport": {
@@ -67,7 +69,17 @@ class TestSessions:
             },
             extension_id="extensionId",
             keep_alive=True,
-            proxies=True,
+            proxies=[
+                {
+                    "type": "browserbase",
+                    "domain_pattern": "domainPattern",
+                    "geolocation": {
+                        "country": "xx",
+                        "city": "city",
+                        "state": "xx",
+                    },
+                }
+            ],
             region="us-west-2",
             api_timeout=60,
             user_metadata={"foo": "bar"},
@@ -143,7 +155,7 @@ class TestSessions:
             project_id="projectId",
             status="REQUEST_RELEASE",
         )
-        assert_matches_type(Session, session, path=["response"])
+        assert_matches_type(SessionUpdateResponse, session, path=["response"])
 
     @parametrize
     def test_raw_response_update(self, client: Browserbase) -> None:
@@ -156,7 +168,7 @@ class TestSessions:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         session = response.parse()
-        assert_matches_type(Session, session, path=["response"])
+        assert_matches_type(SessionUpdateResponse, session, path=["response"])
 
     @parametrize
     def test_streaming_response_update(self, client: Browserbase) -> None:
@@ -169,7 +181,7 @@ class TestSessions:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             session = response.parse()
-            assert_matches_type(Session, session, path=["response"])
+            assert_matches_type(SessionUpdateResponse, session, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -220,7 +232,7 @@ class TestSessions:
         session = client.sessions.debug(
             "id",
         )
-        assert_matches_type(SessionLiveURLs, session, path=["response"])
+        assert_matches_type(SessionDebugResponse, session, path=["response"])
 
     @parametrize
     def test_raw_response_debug(self, client: Browserbase) -> None:
@@ -231,7 +243,7 @@ class TestSessions:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         session = response.parse()
-        assert_matches_type(SessionLiveURLs, session, path=["response"])
+        assert_matches_type(SessionDebugResponse, session, path=["response"])
 
     @parametrize
     def test_streaming_response_debug(self, client: Browserbase) -> None:
@@ -242,7 +254,7 @@ class TestSessions:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             session = response.parse()
-            assert_matches_type(SessionLiveURLs, session, path=["response"])
+            assert_matches_type(SessionDebugResponse, session, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -255,7 +267,9 @@ class TestSessions:
 
 
 class TestAsyncSessions:
-    parametrize = pytest.mark.parametrize("async_client", [False, True], indirect=True, ids=["loose", "strict"])
+    parametrize = pytest.mark.parametrize(
+        "async_client", [False, True, {"http_client": "aiohttp"}], indirect=True, ids=["loose", "strict", "aiohttp"]
+    )
 
     @parametrize
     async def test_method_create(self, async_client: AsyncBrowserbase) -> None:
@@ -291,7 +305,9 @@ class TestAsyncSessions:
                         "min_width": 0,
                     },
                 },
+                "headful": True,
                 "log_session": True,
+                "os": "windows",
                 "record_session": True,
                 "solve_captchas": True,
                 "viewport": {
@@ -301,7 +317,17 @@ class TestAsyncSessions:
             },
             extension_id="extensionId",
             keep_alive=True,
-            proxies=True,
+            proxies=[
+                {
+                    "type": "browserbase",
+                    "domain_pattern": "domainPattern",
+                    "geolocation": {
+                        "country": "xx",
+                        "city": "city",
+                        "state": "xx",
+                    },
+                }
+            ],
             region="us-west-2",
             api_timeout=60,
             user_metadata={"foo": "bar"},
@@ -377,7 +403,7 @@ class TestAsyncSessions:
             project_id="projectId",
             status="REQUEST_RELEASE",
         )
-        assert_matches_type(Session, session, path=["response"])
+        assert_matches_type(SessionUpdateResponse, session, path=["response"])
 
     @parametrize
     async def test_raw_response_update(self, async_client: AsyncBrowserbase) -> None:
@@ -390,7 +416,7 @@ class TestAsyncSessions:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         session = await response.parse()
-        assert_matches_type(Session, session, path=["response"])
+        assert_matches_type(SessionUpdateResponse, session, path=["response"])
 
     @parametrize
     async def test_streaming_response_update(self, async_client: AsyncBrowserbase) -> None:
@@ -403,7 +429,7 @@ class TestAsyncSessions:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             session = await response.parse()
-            assert_matches_type(Session, session, path=["response"])
+            assert_matches_type(SessionUpdateResponse, session, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -454,7 +480,7 @@ class TestAsyncSessions:
         session = await async_client.sessions.debug(
             "id",
         )
-        assert_matches_type(SessionLiveURLs, session, path=["response"])
+        assert_matches_type(SessionDebugResponse, session, path=["response"])
 
     @parametrize
     async def test_raw_response_debug(self, async_client: AsyncBrowserbase) -> None:
@@ -465,7 +491,7 @@ class TestAsyncSessions:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         session = await response.parse()
-        assert_matches_type(SessionLiveURLs, session, path=["response"])
+        assert_matches_type(SessionDebugResponse, session, path=["response"])
 
     @parametrize
     async def test_streaming_response_debug(self, async_client: AsyncBrowserbase) -> None:
@@ -476,7 +502,7 @@ class TestAsyncSessions:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             session = await response.parse()
-            assert_matches_type(SessionLiveURLs, session, path=["response"])
+            assert_matches_type(SessionDebugResponse, session, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
